Introduce step paramterisation.
This is an example of regular expression step matching. This is mainly for
educational purposes. I recommend you use the default parse-based matching
unless it really doesn't work for you.

Note the bolding of the part of the step text used as a parameter when you
run behave. You can try changing the flight number to any other flight in
the database and it should still work.